+++ src/acc/hip-wf-row/wavefront_row_lds.hpp
@@ -15,7 +15,7 @@
 #include "../common/utils.h"
 
 template <unsigned int BLOCK_SIZE, unsigned int WF_SIZE>
-__global__ void device_spmv_wf_row_lds(int trans, const int alpha, const int beta, int m, int n, const int *rowptr,
+__global__ void device_spmv_wf_row_lds(int trans, const double alpha, const double beta, int m, int n, const int *rowptr,
                                        const int *colindex, const double *value, const double *x, double *y) {
   // thread id in block
   int block_thread_id = threadIdx.x;
